finnhub-stats-fetch/requirements.txt
peewee
finnhub-python
websocket-client
pymysql
cryptography